pyproject.toml
[tool.poetry]
name = "book_shop_drf_p23"
version = "0.1.0"
description = ""
authors = ["krv006 <rvkamronbek@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
drf-spectacular = "^0.27.2"
flake8 = "^7.1.1"
isort = "^5.13.2"
python-dotenv = "^1.0.1"
django-mptt = "^0.16.0"
pillow = "^10.4.0"
django-ckeditor-5 = "^0.2.13"
djangorestframework-simplejwt = "^5.3.1"
celery = {extras = ["redis"], version = "^5.4.0"}
django-celery-results = "^2.5.1"
rediscache = "^1.0.0"
django-redis = "^5.4.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
